Refactor transformer engine context manager
